[minor] rename folder
